observability_ai_assistant/requirements.txt
fastmcp
ollama
requests
pydantic
python-dotenv
rich
fastapi
uvicorn
